Separated Features
Broke up all of the features into their respective files, then added in the necessary import statements
