+++ features/makeDir.py
@@ -1,0 +1,3 @@
+def makeDir(sftp, dirName):
+    sftp.mkdir(dirName)
+    print(dirName + " has been made!")
